Use it in interactive mode
