+++ mistralrs-core/src/speech_models/dia/mod.rs
@@ -1,4 +1,4 @@
-use std::{io::Write, sync::Arc};
+use std::sync::Arc;
 
 use audio::{apply_audio_delay, build_delay_indices, build_revert_indices, revert_audio_delay};
 use cache::DiaKvCache;
@@ -14,91 +14,13 @@ use rand_isaac::Isaac64Rng;
 
 pub use config::DiaConfig;
 
-use super::bs1770;
+use super::utils::normalize_loudness;
 
 mod audio;
 mod cache;
 mod config;
 mod dac;
 mod model;
-
-fn normalize_loudness(wav: &Tensor, sample_rate: u32, loudness_compressor: bool) -> Result<Tensor> {
-    let energy = wav.sqr()?.mean_all()?.sqrt()?.to_vec0::<f32>()?;
-    if energy < 2e-3 {
-        return Ok(wav.clone());
-    }
-    let wav_array = wav.to_vec1::<f32>()?;
-    let mut meter = bs1770::ChannelLoudnessMeter::new(sample_rate);
-    meter.push(wav_array.into_iter());
-    let power = meter.as_100ms_windows();
-    let loudness = match bs1770::gated_mean(power) {
-        None => return Ok(wav.clone()),
-        Some(gp) => gp.loudness_lkfs() as f64,
-    };
-    let delta_loudness = -14. - loudness;
-    let gain = 10f64.powf(delta_loudness / 20.);
-    let wav = (wav * gain)?;
-    if loudness_compressor {
-        wav.tanh()
-    } else {
-        Ok(wav)
-    }
-}
-
-pub trait Sample {
-    fn to_i16(&self) -> i16;
-}
-
-impl Sample for f32 {
-    fn to_i16(&self) -> i16 {
-        (self.clamp(-1.0, 1.0) * 32767.0) as i16
-    }
-}
-
-impl Sample for f64 {
-    fn to_i16(&self) -> i16 {
-        (self.clamp(-1.0, 1.0) * 32767.0) as i16
-    }
-}
-
-impl Sample for i16 {
-    fn to_i16(&self) -> i16 {
-        *self
-    }
-}
-
-fn write_pcm_as_wav<W: Write, S: Sample>(
-    w: &mut W,
-    samples: &[S],
-    sample_rate: u32,
-) -> std::io::Result<()> {
-    let len = 12u32; // header
-    let len = len + 24u32; // fmt
-    let len = len + samples.len() as u32 * 2 + 8; // data
-    let n_channels = 1u16;
-    let bytes_per_second = sample_rate * 2 * n_channels as u32;
-    w.write_all(b"RIFF")?;
-    w.write_all(&(len - 8).to_le_bytes())?; // total length minus 8 bytes
-    w.write_all(b"WAVE")?;
-
-    // Format block
-    w.write_all(b"fmt ")?;
-    w.write_all(&16u32.to_le_bytes())?; // block len minus 8 bytes
-    w.write_all(&1u16.to_le_bytes())?; // PCM
-    w.write_all(&n_channels.to_le_bytes())?; // one channel
-    w.write_all(&sample_rate.to_le_bytes())?;
-    w.write_all(&bytes_per_second.to_le_bytes())?;
-    w.write_all(&2u16.to_le_bytes())?; // 2 bytes of data per sample
-    w.write_all(&16u16.to_le_bytes())?; // bits per sample
-
-    // Data block
-    w.write_all(b"data")?;
-    w.write_all(&(samples.len() as u32 * 2).to_le_bytes())?;
-    for sample in samples.iter() {
-        w.write_all(&sample.to_i16().to_le_bytes())?
-    }
-    Ok(())
-}
 
 fn create_attn_mask(q_padding_mask_1d: &Tensor, k_padding_mask_1d: &Tensor) -> Result<Tensor> {
     let (b1, _tq) = q_padding_mask_1d.dims2()?;
