+++ mistralrs-core/src/speech_models/mod.rs
@@ -1,4 +1,5 @@
 mod bs1770;
 mod dia;
+pub mod utils;
 
 pub use dia::{DiaConfig, DiaPipeline};
